BSD License New header Replacement
Adding in new BSD License on all the files in the library, Also fixed
any copyright naming in the files since now that top name and copyright
directly relate to the BSD License wording. Added in the start of what
will go into the xml nodes as a header as well in the Library renamer
xml file.