Region usuabiltiy work
- converted regions to a left to right workflow.
This seems to work better with the input channels being top to bottom.
It also has the benefit of automatic input to output handle targeting
without the middle step of the handle selection dialog.

- UseRegion is no longer an if statement but a value that gets
multiplied against the mask. The skipped if statement seems to gain us a
few instructions on the gpu.