ChannelLayerMask: Lots of Logic added
If the selected layer does not contain a mask or maskstack and only one
channel layer is added a maskstack is created and
the channel layer is added to the maskstack

If the selected layer does not contain a mask or maskstack and multiple
channel layers are selected the first added channel
layer in the mask stack will be set to Blendmode 'Normal', subsequent
ones to Blendmode 'Screen'

If the selected layer does contain a mask or maskstack, channel layers
will be added to the existing mask/maskstack.

If the selected layer does contain a mask or maskstack and only one
channel layer is added, it will be set to Blendmode 'Multiply'.
If 'Invert' was on during Channel Layer creation, the channel layer and
invert adjustment will be grouped to prevent
existing maskdata from being inverted.

If the selected layer does contain a mask or maskstack  and multiple
channel layers are added, they will be grouped first
with the group's Blendmode set to 'Multiply'. ChannelLayers will then be
added to the Group with the first one at
Blendmode 'Normal', subsequent ones at Blendmode 'Screen'.