Added new options to 'Tool/Transform Paint' Tool Properties Toolbar
New options were added to the Tool Properties Toolbar of the
TransformPaint Tool:

- Buffer Resolution:
Gives you the option to change the buffer resolution.
Inherited and connected to Painting Palette.

- Buffer Res Double/Halve
Convenience Buttons to quickly double or halve the buffer resolution

- Buffer Depth
Allows you to set the Buffer Bitdepth.
Linked to Painting Palette. Please note, there is a known signal bug in
Mari
that prevents the dropdown to react to changes made directly from the
painting palette. It will only update once you switch tools.

- Clamp
Allows you to toggle the clamping of the paintbuffer.
Linked to Painting Palette.
Please note there is no signal to live update this checkbox if the
checkbox in the Painting Palette is used. You need to switch tools for
the checkbox to update to any external changes.

- Sync Buffer BitDepth
When checked your Buffer BitDepth will be kept in sync with the Bitdepth
of the currently active channel.
This setting is saved and restored across projects.

- Flip X/Y
Flips the Paintbuffer horizontally or vertically